Make V₁_basis_matrix into an orthogonal basis, but not a basis over ℤ anymore.
Originally (see B&P), you consider V₁ := v₀^⟂ and find a ℤ-basis for
V₁, and a set W of representatives for the sublattice v₀(+)V₁.
The problem is that our basis here is not orthogonal, and makes all
computations more complicated.

The goal now is to actually diagonalize first the form, and thus get a
basis for V₁ _over ℚ_ but not over ℤ anymore.
Since the code as it was written until now didn't depend on having a
basis over ℤ, we're safe I think.
The next steps are to use this during root enumeration to reduce the
number of roots to try.